# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manu local HTTP API",
    "version": "v1",
    "description": "Contrato HTTP local, sem autenticação, para o primeiro recorte de ingestão, consulta e inspeção de evidências. A instância é limitada ao endereço loopback e usa uma única organização configurada; o valor padrão é `local`. Os envelopes JSON têm versão `v1`. Erros usam `application/problem+json` e não expõem conteúdo bruto, segredos ou diagnósticos de provedores."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Servidor local padrão (o endereço configurado deve permanecer em loopback)."
    }
  ],
...
  "paths": {
...
    "/api/v1/queries": {
      "post": {
        "tags": ["queries"],
        "summary": "Executa uma consulta de forma síncrona",
        "description": "O corpo aceita somente JSON conhecido e exige o tipo epistemológico `kind`. A execução é síncrona e retorna em estado terminal: `completed`, `partial`, `abstained` ou `failed`. Quando há resultado, `claims[].id` é estável e cada citação aponta para uma Evidence Unit reinspecionável; a política local pode produzir abstinência determinística antes de qualquer provedor.",
        "operationId": "createQuery",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/QueryRequest"},
              "example": {
                "question": "Quais fluxos são observados neste sistema?",
                "kind": "possible_flow",
                "source_id": "22222222-2222-4222-8222-222222222222",
                "snapshot_id": "33333333-3333-4333-8333-333333333333"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consulta concluída em `completed`, `partial` ou `abstained`; `failed` usa 500 com o mesmo envelope de consulta.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QueryResponse"}}}
          },
          "400": {"$ref": "#/components/responses/InvalidQuery"},
          "408": {"$ref": "#/components/responses/RequestTimeout"},
          "409": {"$ref": "#/components/responses/QueryConflict"},
          "413": {"$ref": "#/components/responses/QueryLimit"},
          "415": {"$ref": "#/components/responses/UnsupportedMedia"},
          "500": {
            "description": "Falha da execução ou resultado terminal `failed`. Falhas do serviço usam problem+json; uma execução `failed` persistida usa o envelope QueryResponse.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {
              "application/json": {"schema": {"$ref": "#/components/schemas/QueryResponse"}},
              "application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}
            }
          },
          "503": {"$ref": "#/components/responses/QueryUnavailable"},
          "405": {"$ref": "#/components/responses/MethodNotAllowedPost"}
        }
      }
    },
    "/api/v1/queries/{query_id}": {
      "parameters": [{"$ref": "#/components/parameters/QueryID"}],
      "get": {
        "tags": ["queries"],
        "summary": "Reinspeciona uma execução de consulta",
        "description": "Permite observar estados `pending` e `running` persistidos, além dos estados terminais. O identificador deve ser um UUID textual de 36 caracteres.",
        "operationId": "getQuery",
        "responses": {
          "200": {
            "description": "Envelope da execução.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QueryResponse"}}}
          },
          "400": {"$ref": "#/components/responses/InvalidQueryPath"},
          "404": {"$ref": "#/components/responses/QueryNotFound"},
          "408": {"$ref": "#/components/responses/RequestTimeout"},
          "500": {"$ref": "#/components/responses/QueryFailed"},
          "503": {"$ref": "#/components/responses/QueryUnavailable"},
          "405": {"$ref": "#/components/responses/MethodNotAllowedGet"}
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "RequestID": {
        "description": "Identificador de correlação ecoado no corpo e na resposta. Se o request não o fornecer, o servidor gera um valor seguro.",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9_.-]{1,128}$"
        },
        "example": "req_example_01"
      }
    },
    "parameters": {
...
      "QueryID": {
        "name": "query_id",
        "in": "path",
        "required": true,
        "description": "Identificador UUID textual da consulta.",
        "schema": {"$ref": "#/components/schemas/UUID"},
        "example": "66666666-6666-4666-8666-666666666666"
      },
...
    },
    "responses": {
...
      "InvalidQuery": {"$ref": "#/components/responses/Problem400Query"},
      "InvalidQueryPath": {"$ref": "#/components/responses/Problem400QueryPath"},
      "QueryNotFound": {"$ref": "#/components/responses/Problem404Query"},
      "QueryConflict": {"$ref": "#/components/responses/Problem409Query"},
      "QueryLimit": {"$ref": "#/components/responses/Problem413Query"},
      "QueryFailed": {"$ref": "#/components/responses/Problem500Query"},
      "QueryUnavailable": {"$ref": "#/components/responses/Problem503Query"},
...
      "RequestTimeout": {
        "description": "O request foi cancelado ou excedeu o prazo; `code` é `request_canceled` ou `request_timeout`.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "request_timeout", "message": "request timed out", "status": 408, "request_id": "req_example_30"}}}
      },
      "UnsupportedMedia": {
        "description": "O media type não é aceito pela operação.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "unsupported_media_type", "message": "content type is not supported", "status": 415, "request_id": "req_example_31"}}}
      },
      "MethodNotAllowedPost": {
        "description": "O método não é permitido para a coleção; `Allow: POST`.",
        "headers": {"Allow": {"schema": {"type": "string", "const": "POST"}}, "X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "method_not_allowed", "message": "method is not allowed", "status": 405, "request_id": "req_example_32"}}}
      },
      "MethodNotAllowedGet": {
        "description": "O método não é permitido para o recurso; `Allow: GET`.",
        "headers": {"Allow": {"schema": {"type": "string", "const": "GET"}}, "X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "method_not_allowed", "message": "method is not allowed", "status": 405, "request_id": "req_example_33"}}}
      },
...
      "Problem400Query": {
        "description": "Request de consulta inválido (`invalid_query`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "invalid_query", "message": "query request is invalid", "status": 400, "request_id": "req_example_41"}}}
      },
      "Problem400QueryPath": {
        "description": "O identificador ou caminho da consulta é inválido (`invalid_query_path`); um resultado persistido inválido pode usar `invalid_query`.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "invalid_query_path", "message": "query path is invalid", "status": 400, "request_id": "req_example_42"}}}
      },
      "Problem404Query": {
        "description": "Consulta não encontrada (`query_not_found`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "query_not_found", "message": "query was not found", "status": 404, "request_id": "req_example_43"}}}
      },
      "Problem409Query": {
        "description": "Conflito de execução da consulta (`query_conflict`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "query_conflict", "message": "query conflicts with existing execution", "status": 409, "request_id": "req_example_44"}}}
      },
      "Problem413Query": {
        "description": "Limite do corpo da consulta excedido (`query_limit_exceeded`); uma rejeição global anterior ao roteamento pode usar `body_too_large`. O padrão é 16 KiB, configurável.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "query_limit_exceeded", "message": "query exceeds configured limits", "status": 413, "request_id": "req_example_45"}}}
      },
      "Problem500Query": {
        "description": "Falha da consulta (`query_failed`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "query_failed", "message": "query failed", "status": 500, "request_id": "req_example_46"}}}
      },
      "Problem503Query": {
        "description": "Consulta indisponível, ocupada ou ainda não terminou (`query_unavailable`, `query_busy` ou `query_not_complete`); o limite global de concorrência pode usar `concurrency_limited`.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "query_unavailable", "message": "query service is unavailable", "status": 503, "request_id": "req_example_47"}}}
      },
...
    },
    "schemas": {
      "UUID": {
        "type": "string",
        "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
        "description": "UUID textual de 36 caracteres; a validação do handler aceita maiúsculas e minúsculas."
      },
      "Digest": {
        "type": "string",
        "pattern": "^[0-9a-f]{64}$",
        "description": "Digest SHA-256 em hexadecimal minúsculo."
      },
      "RequestID": {
        "type": "string",
        "minLength": 1,
        "maxLength": 128,
        "pattern": "^[A-Za-z0-9_.-]{1,128}$"
      },
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "code", "message", "status", "request_id"],
        "properties": {
          "version": {"type": "string", "const": "v1"},
          "code": {"type": "string", "pattern": "^[a-z0-9_.-]{1,64}$"},
          "message": {"type": "string", "maxLength": 128},
          "status": {"type": "integer", "minimum": 400, "maximum": 599},
          "request_id": {"$ref": "#/components/schemas/RequestID"}
        },
        "description": "Envelope de erro seguro. O texto é fixo e não carrega conteúdo do request, DSN, segredo ou mensagem de provedor."
      },
...
      "QueryRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["question", "kind"],
        "properties": {
          "question": {"type": "string", "minLength": 1, "maxLength": 16384, "description": "Pergunta não vazia, sem NUL, CR ou LF; o limite padrão do corpo é 16 KiB."},
          "kind": {"type": "string", "enum": ["inventory", "possible_flow", "observed_execution", "business_intent"], "description": "Tipo epistemológico declarado pelo cliente; não é inferido do texto da pergunta."},
          "source_id": {"$ref": "#/components/schemas/UUID"},
          "snapshot_id": {"$ref": "#/components/schemas/UUID", "description": "Opcional; quando informado, `source_id` também é obrigatório."}
        },
        "description": "A organização é fixa na configuração e não pode ser enviada como campo ou header de escopo."
      },
      "QueryResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "id", "organization_id", "state", "question_digest", "request_id"],
        "properties": {
          "version": {"type": "string", "const": "v1"},
          "id": {"$ref": "#/components/schemas/UUID"},
          "organization_id": {"type": "string", "minLength": 1},
          "source_id": {"$ref": "#/components/schemas/UUID"},
          "snapshot_id": {"$ref": "#/components/schemas/UUID"},
          "state": {"type": "string", "enum": ["pending", "running", "completed", "partial", "failed", "abstained"]},
          "question_digest": {"$ref": "#/components/schemas/Digest"},
          "package_digest": {"$ref": "#/components/schemas/Digest"},
          "result": {"$ref": "#/components/schemas/QueryResult"},
          "diagnostic_code": {"type": "string", "pattern": "^[a-z0-9_.-]{1,64}$"},
          "created_at": {"type": "string", "format": "date-time"},
          "started_at": {"type": "string", "format": "date-time"},
          "finished_at": {"type": "string", "format": "date-time"},
          "request_id": {"$ref": "#/components/schemas/RequestID"}
        },
        "description": "Envelope de status. `partial` e `abstained` são estados observáveis; não significam que uma hipótese foi promovida a execução observada."
      },
      "QueryResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "knowledge_state", "text", "claims", "citations", "gaps", "generation"],
        "properties": {
          "version": {"type": "string", "const": "v1alpha1"},
          "knowledge_state": {"type": "string", "enum": ["generated_reviewable"]},
          "text": {"type": "string", "maxLength": 65536},
          "claims": {"type": "array", "items": {"$ref": "#/components/schemas/Claim"}},
          "citations": {"type": "array", "items": {"$ref": "#/components/schemas/Citation"}},
          "gaps": {"type": "array", "items": {"$ref": "#/components/schemas/Gap"}},
          "generation": {"$ref": "#/components/schemas/GenerationMetadata"}
        },
        "description": "Resultado gerado e revisável. Cada afirmação possui identificador estável e suas citações apontam para Evidence Units reinspecionáveis."
      },
      "Claim": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "ordinal", "kind", "support", "text"],
        "properties": {
          "id": {"$ref": "#/components/schemas/UUID"},
          "ordinal": {"type": "integer", "minimum": 1},
          "kind": {"type": "string", "enum": ["observed", "generated", "gap"]},
          "support": {"type": "string", "enum": ["supported", "unsupported", "abstained"]},
          "text": {"type": "string", "maxLength": 16384},
          "citation_ordinals": {"type": "array", "items": {"type": "integer", "minimum": 1}},
          "gap_ordinals": {"type": "array", "items": {"type": "integer", "minimum": 1}}
        }
      },
      "Citation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["ordinal", "organization_id", "source_id", "snapshot_id", "evidence_id", "locator", "role"],
        "properties": {
          "ordinal": {"type": "integer", "minimum": 1},
          "organization_id": {"$ref": "#/components/schemas/UUID"},
          "source_id": {"$ref": "#/components/schemas/UUID"},
          "snapshot_id": {"$ref": "#/components/schemas/UUID"},
          "evidence_id": {"type": "string", "minLength": 1, "maxLength": 256},
          "locator": {"$ref": "#/components/schemas/Locator"},
          "role": {"type": "string", "enum": ["supports", "contests", "context"]}
        }
      },
      "Gap": {
        "type": "object",
        "additionalProperties": false,
        "required": ["ordinal", "id", "code", "message"],
        "properties": {
          "ordinal": {"type": "integer", "minimum": 1},
          "id": {"type": "string", "minLength": 1, "maxLength": 256},
          "code": {"type": "string", "pattern": "^[a-z0-9_.-]{1,64}$"},
          "message": {"type": "string", "maxLength": 4096}
        }
      },
      "GenerationMetadata": {
        "type": "object",
        "additionalProperties": false,
        "required": ["provider", "model", "profile", "protocol", "usage", "termination", "package_id", "package_digest", "query_id", "query_digest", "started_at", "finished_at", "latency"],
        "properties": {
          "provider": {"type": "string", "enum": ["none", "simulated", "openai", "openai-compatible", "openrouter"]},
          "model": {"type": "string", "maxLength": 128},
          "profile": {"type": "string", "maxLength": 128},
          "protocol": {"type": "string", "enum": ["none", "responses", "chat_completions"]},
          "usage": {"type": "object", "additionalProperties": false, "required": ["input_items", "output_items", "input_tokens", "output_tokens"], "properties": {"input_items": {"type": "integer", "minimum": 0}, "output_items": {"type": "integer", "minimum": 0}, "input_tokens": {"type": "integer", "minimum": 0}, "output_tokens": {"type": "integer", "minimum": 0}}},
          "termination": {"type": "string", "enum": ["completed", "partial", "abstained"]},
          "package_id": {"type": "string", "minLength": 1, "maxLength": 256},
          "package_digest": {"$ref": "#/components/schemas/Digest"},
          "query_id": {"type": "string", "minLength": 1, "maxLength": 256},
          "query_digest": {"$ref": "#/components/schemas/Digest"},
          "started_at": {"type": "string", "format": "date-time"},
          "finished_at": {"type": "string", "format": "date-time"},
          "latency": {"type": "integer", "minimum": 0}
        }
      },
...
    }
  }
}